+++ src/Lesson1/Mission.java
@@ -1,0 +1,9 @@
+package Lesson1;
+
+public interface Mission {
+
+    boolean doRun(RunAndJump runAndJump);
+    boolean doJump(RunAndJump runAndJump);
+    boolean isTrack();
+    String getTitle();
+}
