+++ src/Lesson1/RunAndJump.java
@@ -1,0 +1,10 @@
+package Lesson1;
+
+public interface RunAndJump {
+
+    void run();     //Для первого задания
+    void jump();    //Для первого задания
+    int getRun();
+    int getJump();
+    String getName();
+}
