--- src/Mission.java
@@ -1,7 +1,0 @@
-public interface Mission {
-
-    boolean doRun(RunAndJump runAndJump);
-    boolean doJump(RunAndJump runAndJump);
-    boolean isTrack();
-    String getTitle();
-}
