--- src/RunAndJump.java
@@ -1,8 +1,0 @@
-public interface RunAndJump {
-
-    void run();     //Для первого задания
-    void jump();    //Для первого задания
-    int getRun();
-    int getJump();
-    String getName();
-}
